pyproject.toml
[build-system]  # comment
requires = ["setuptools>=65", "wheel"]  # comment
build-backend = "setuptools.build_meta"  # comment

[project]  # comment
name = "finrobot-pixiu-eval"  # comment
version = "0.1.0"  # comment
description = "Minimal reproducible FinRobot agent evaluation on PIXIU/FinBen"  # comment
readme = "README.md"  # comment
requires-python = ">=3.10"  # comment
license = {text = "MIT"}  # comment
authors = [{name = "CS263 Project"}]  # comment

dependencies = [  # comment
  "openai>=1.0.0",  # comment
  "pyautogen[retrievechat]==0.2.19",  # comment
  "datasets>=2.18.0",  # comment
  "pandas>=1.5.0",  # comment
  "numpy>=1.23.0",  # comment
  "tqdm>=4.66.0",  # comment
  "pyyaml>=6.0",  # comment
  "rich>=13.0",  # comment
  "requests>=2.31.0",  # comment
  "ipython>=8.0.0",  # comment
  "typing_extensions>=4.0.0",  # comment
  "finnhub-python>=2.4.0",  # comment
  "yfinance>=0.2.28",  # comment
  "sec-api>=1.0.0",  # comment
  "praw>=7.7.0",  # comment
  "mplfinance>=0.12.9b7",  # comment
  "matplotlib>=3.7.0",  # comment
  "backtrader>=1.9.78",  # comment
  "reportlab>=4.0.0",  # comment
  "PyPDF2>=3.0.0",  # comment
  "pytest>=7.0.0"  # comment
]  # comment
  # comment
[project.optional-dependencies]  # comment
full = [  # comment
  "finnlp",  # comment
  "pandas_datareader>=0.10.0",  # comment
  "tushare>=1.2.89",  # comment
  "pdfkit==1.0.0",  # comment
  "marker-pdf",  # comment
  "unstructured==0.8.1",  # comment
  "nltk==3.8.1",  # comment
  "scikit_learn==1.5.0",  # comment
  "ratelimit==2.2.1",  # comment
  "tenacity==8.3.0",  # comment
  "aiohttp==3.8.5"  # comment
]  # comment

[tool.setuptools]  # comment
package-dir = {"" = "src"}  # comment

[tool.setuptools.packages.find]  # comment
where = ["src"]  # comment
